# [Xamarin.Android.Build.Tasks] Fix up ZipArchiveEx to handle updates.