--- a/Documentation/release-notes/5266.md
+++ b/Documentation/release-notes/5266.md
@@ -1,0 +1,6 @@
+### Build and deployment performance
+
+  * [GitHub PR 5266](https://github.com/xamarin/xamarin-android/pull/5266):
+  Minor improvements to the way zip archives are created. Do not create
+  folders in the classes.zip file and allow the auto flushing of
+  a zip file to be switched off.