--- a/src/Xamarin.Android.Build.Tasks/Tasks/Javac.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/Javac.cs
@@ -39,7 +39,6 @@ namespace Xamarin.Android.Tasks
 			if (!string.IsNullOrEmpty (ClassesZip)) {
 				using (var zip = new ZipArchiveEx (ClassesZip, FileMode.OpenOrCreate)) {
 					zip.AutoFlush = false;
-					zip.CreateDirectoriesInZip = false;
 					zip.AddDirectory (ClassesOutputDirectory, "", CompressionMethod.Store);
 				}
 			}