While investigating another bug, I found that the code to look up
an existing entry in the zip was not working at all.

The problem was the filename we were looking up included a `/` at
the start of the path. The zip entires in the zip do not include a
`/` at the start. So the code was never finding a match. As a result
we were always trying to add a file rather than skipping those which
were up to date.

In addition I found that we do not always need to create entires in
the zip for directories. So when we create the `classes.zip` in the
`Javac` task we dont need to create directory entires. So we should
have a property to allow us to skip directories.

Also we should have the ability to turn the auto flush on/off.

There seems to be a small performance boost with these changes.

Old
```
749 ms  Javac                     
746 ms  Javac                   
```

New
```
728 ms  Javac 
724 ms  Javac 
```