@jonpryor 2020-12-09
@dellis1972 : please review some of the unit test failures: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=4286406&view=ms.vss-test-web.build-test-results-tab

e.g. https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=4286406&view=ms.vss-test-web.build-test-results-tab&runId=17549180&resultId=100251&paneView=debug

```
Expected string length 31 but was 32. Strings differ at index 18.
Expected: "temp/A.txt\ntemp/B/\ntemp/B/B.txt"
But was:  "temp/A.txt\ntemp/B//\ntemp/B/B.txt"
------------------------------^
```