@dellis1972 2020-12-17
@jonpryor I will add a size limit , but I need to keep the file count limit as well because of this error

```
Xamarin.Tools.Zip.ZipException: Failure to create temporary file: Too many open files
```

which is :'(